now we are using instancing differently in voxel drawing.

+++ GameApi/GameApi_h.hh
@@ -567,7 +567,7 @@ struct PointArray2
 struct PointArray3
 {
   float *array;
-  unsigned int *color=0;
+  float *color=0; // 4 items per numpoint
   Vector *normal=0;
   int numpoints;
   Low_GLuint buffer[2];
@@ -578,7 +578,7 @@ struct PointArray3
 struct MatrixArray3
 {
   float *array;
-  unsigned int *color=0;
+  float *color=0; // 4 items per numpoint
   Vector *normal=0;
   int numpoints;
   Low_GLuint buffer[2];
